src/ai/model.py
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
import numpy as np
import pandas as pd

class AIModel:
    def __init__(self, data):
        self.data = data
        self.model = LinearRegression()
        self.trained = False

    def preprocess_data(self):
        # Example preprocessing: fill missing values and normalize
        self.data.fillna(self.data.mean(), inplace=True)
        self.data = (self.data - self.data.mean()) / self.data.std()

    def train(self, target_column):
        self.preprocess_data()
        X = self.data.drop(columns=[target_column])
        y = self.data[target_column]
        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
        self.model.fit(X_train, y_train)
        self.trained = True
        return self.model.score(X_test, y_test)

    def predict(self, input_data):
        if not self.trained:
            raise Exception("Model is not trained yet.")
        return self.model.predict(np.array(input_data).reshape(1, -1))

    def load_prompts(self, prompt_file):
        with open(prompt_file, 'r') as file:
            self.prompts = file.readlines()

    def generate_response(self, input_text):
        # Placeholder for response generation logic
        return f"Response based on input: {input_text}"